Cosmetics and frame_id management in TPC:
Seperate correctly with 2 Chase Pyndiah decoder and sub encoder/decoder for rows and cols.
Force to use 1 inter frame level, because it does not work.
Develop the use of the frame_id to manage automatically the address offsets in vectors. Sub column encoder-decoder have an inter frame level equals to the sub row encoder-decoder width + parity bit if any, and reciproquelly.