Fix err tracker:
Add an attribute to facotry modem parameters that has the channel type.
Build the modem (OOK, Optical) in function of it instead of a parameter in the build function.
Fix then the use of USER channel in err tracker revert with OOK and Optical channels because modems weren't built correctly.